Add 'isFilled' field to Polygon class, fix issue with doted border draws
The new field 'isFilled' in Polygon class allow you to chose if you want the Polygon to be filled when drawn. When I have adding this I also found that there was an issue with doted borders drawing. The last border was missing, I fixed it (_paintDottedLine method). Also added an example screen for the Polygon Layer.