fix ; mapController.moveAndRotate with null id
The id variable in the mapController moveAndRotate function is declared as nullable, but a null check is applied. This causes the function to not work if it is called without the optional id field.

Suggest casting the String? to a String type to allow the null id value to be used.