Polygon Improvements (#501)
* Add 'isFilled' field to Polygon class, fix issue with doted border draws

The new field 'isFilled' in Polygon class allow you to chose if you want the Polygon to be filled when drawn. When I have adding this I also found that there was an issue with doted borders drawing. The last border was missing, I fixed it (_paintDottedLine method). Also added an example screen for the Polygon Layer.

* Reformat files

* Sort imports + fix duplicated imports

* Run dark format . --fix

* Fix exmaple/lib/pages/Polygon.dart imports and upgraded to latlong2 package and other minor fixes

* update polygon_layer.dart to handle isFilled arg

* Removed ; lib/src/layer/polygon_layer.dart:253