Fix Tile tileReady callback race condition
The Tile loading was initiated before adding the Tile to the TileManager
which meant that with a very fast connection or cached Tile images the
load callback would be called before the Tile was in the TileManager.
Therefore when we tried to set the loaded value of the Tile in the
callback we would not find the Tile and the Tile was never marked as
loaded.

This caused the tile to be removed sooner than it should be since we
prune tiles that are not loaded yet when zooming in.