Prevent double scrolling
Before this change the PointerScrollEvent handler was not registered
with PointerSignalResolver which meant that if scrolling was enabled for
FlutterMap and it was within a scrollable widget or had a scrollable
child widget then they would both be scrolled simultaneously. This is
only noticeable on web since on mobile a drag gesture is used to
scroll/pan instead of mouse scrolling.