Replace TileLayer's error image option with a placeholder image option
When a tile fails to load and there are no existing loaded tiles which
overlay with it from another zoom level we are left with a blank portion
on the map. When there are no loaded tiles visible, for example because
of a network problem, there is no visual feedback when
zooming/panning/rotating because the map is just a single solid color.
This prompted me to add a placeholder image.

We already have an error image implementation but the error image will
not show up until loading fails which, in the case of a poor network
connection, can take quite a while.

The placeholder implementation gives visual feedback both whilst loading
and when loading fails.

Note that the current implementation just fills the viewport with the
placeholder tile behind all other tiles. It would be more efficient to
avoid adding placeholder tiles which are completely obscured by other
visual tiles but my attempts to do this were unsuccessful due to the
complex nature of tile pruning.