Replace `CustomPoint` with extension methods on `Point` (#1585)
* Replace CustomPoint with extension methods on Point

The CustomPoint class predates dart extension methods. The name is
confusing (custom in what sense, and why) and in looking at the
implementation it became clear that it was essentially just adding
extension methods.

Changing to a Point will avoid conversions in plugins which deal with
libraries that use Point.

Code will work as-is without changes with the exception of the
subtraction operator. Previously CustomPoint<int> - CustomPoint<double>
did not trigger analyzer warnings but would cause a runtime exception.
This is now replaced by subtract() extension methods which allow
subtracting num from a Point<double> and int from Point<int>.

* Avoid redundant maths imports

* Add Point.toOffset extension method

Co-authored-by: 6y <tlserver6y@gmail.com>

* Use point extension's toOffset()

---------

Co-authored-by: 6y <tlserver6y@gmail.com>