Only create placeholder tiles where tiles are missing or not fully transitioned at the current zoom
This reduces the number of placeholder widgets. The algorithm used to
determine which coordinates to show placeholders for is a tradeoff
between speed and the number of redundant placeholders created. It
creates a placeholder for every tile at the current zoom which failed to
load or is still transitioning. This means that if a tile from a lower
zoom obscures the placeholder or multiple tiles from a higher zoom
collectively obscure the placeholder it will be unnecessarily created.
It would be possible to avoid this but it would require a more complex
data structure or iterating all of the tiles for every potential
placeholder.