Fix crash on uncaught "Invalid image data" exception in case of zero bytes (MacOS desktop).
When running a custom tile server, in the case of a missing tile we can get
back zero bytes. Resulting in the **decode** function failure in
**_loadAsync** of FlutterMapNetworkImageProvider.

Looking at the **_loadAsync** implementation, the call to the **decode**
function was left out of the try/catch block. Simply moving the
**decode** function call inside the try block fixes the issue - the app
no longer crashes on a missing tile.