Ensure the RichAttributionWidgetState doesn't fail if its removed from the render tree.
While there was an assertion to check for this, it being removed ~1-2 frames after its added is perfectly valid, and thus we should easily check if the widget is still mounted.
While we are at it, it seemed a GlobalKey was being used to track the context. In later versions of flutter Stateful widgets have a context property for this purpose.