Simplifed building the RichAttributionWidget.
Previous the widget waited two frames to find the dimensions it was allowed, then set the minWidth to be the width it naturally wanted. I think this was to ensure the widget took
the full width of the screen. It was simple to set a Container's width to infinity to achive the same thing, with less complexity.