Make isPointInPolygon 40% faster (at least in JIT mode).
I mostly looked at the function because I wanted to use it myself but it
was private. So I thought, if I expose it then I could at least put a
ribbon on it by adding a benchmark.

Before:

  (duration: 0:00:05.998949, name: In circle)
  (duration: 0:00:06.866919, name: Not in circle)

After:

  (duration: 0:00:03.649496, name: In circle)
  (duration: 0:00:04.611599, name: Not in circle)

Note, I opportunistically touched crs to remove the dart:ui dependency.
This way it can be compiled with dart (w/o flutter) rendering the
instructions in the benchmark correct again. Unfortunately,
pointInPolygon is not so fortunate and needs flutter due to
the dependency on ui.Offset. That's why I could only run it with
"flutter test" in JIT mode.